# signteb-license-manager/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SignTeb License Manager API
  version: 1.0.0
  description: |
    Licensing, update-delivery and AI service API for all SignTeb products
    (MEDORA AI, SignBot, SignTeb SEO Dashboard, QRCODR, …).

    **Response signing** — every successful `/plugin/*` response is wrapped in an
    envelope `{data, timestamp, signature}` where `signature =
    HMAC-SHA256(json(data) + "|" + timestamp, product_signing_secret)`. The SDK
    must verify before trusting the payload.

    **Rate limits** — public endpoints: 30 req/min/IP; validate: 12 req/min/license;
    AI: 60 req/min/license. Exceeding returns `429` with `Retry-After`.
  contact:
    name: SignTeb
    url: https://signteb.com
servers:
  - url: https://license.signteb.com/api
paths:
...
  /v1/plugin/validate:
    post:
      summary: Validate a license for a domain (SDK heartbeat)
      operationId: validateLicense
      tags: [Plugin Licensing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [license_key, domain, product]
              properties:
                license_key: { type: string }
                domain: { type: string }
                product: { type: string }
      responses:
        '200':
          description: Signed validation result (valid may be false)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedValidation' }
        '404': { $ref: '#/components/responses/InvalidLicense' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /v1/plugin/check-update:
    get:
      summary: Check for a newer product version
      operationId: checkUpdate
      tags: [Updates]
      parameters:
        - { name: license_key, in: query, required: true, schema: { type: string } }
        - { name: domain, in: query, required: true, schema: { type: string } }
        - { name: product, in: query, required: true, schema: { type: string } }
        - { name: version, in: query, required: true, schema: { type: string }, example: 1.2.0 }
        - { name: channel, in: query, schema: { type: string, enum: [stable, beta], default: stable } }
      responses:
        '200':
          description: Signed update manifest
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      update_available: { type: boolean }
                      version: { type: string }
                      release_notes: { type: string }
                      requires_php: { type: string }
                      requires_wp: { type: string }
                      sha256: { type: string, description: SHA-256 of the package zip }
                      download_url: { type: string, format: uri, description: Signed URL, expires in 5 minutes }
                  timestamp: { type: integer }
                  signature: { type: string }
...
components:
  securitySchemes:
    licenseKey:
      type: apiKey
      in: header
      name: X-SLM-License
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
    Validation:
      type: object
      properties:
        valid: { type: boolean }
        status: { type: string, enum: [pending, active, grace, expired, suspended, revoked] }
        grace: { type: boolean, description: True while in the soft-lock renewal window }
        reason: { type: string, nullable: true, example: domain_not_activated }
        expires_at: { type: string, format: date-time, nullable: true }
        grace_ends_at: { type: string, format: date-time, nullable: true }
        activations_used: { type: integer }
        activation_limit: { type: integer }
        product: { type: string }
        plan: { type: string, nullable: true }
    SignedValidation:
      type: object
      properties:
        data: { $ref: '#/components/schemas/Validation' }
        timestamp: { type: integer, description: Unix epoch seconds }
        signature: { type: string, description: HMAC-SHA256 over json(data)|timestamp }
  responses:
    InvalidLicense:
      description: License key not found for this product
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: invalid_license, message: License key not found for this product. }
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema: { type: integer }